KZOO-46: Add Emergency Notification for E911 Compliance (#6443)
This PR adds new functionality to help facilitate compliance with
Kari's Law and notification for when an emergency call is placed.

An email contact list can now be set on a phone numbers emergency
metadata.

Notification can now be achieved one of several ways:

* A new amqp event, `emergency_bridge` has been added to
`kapi_notifications` to trigger a teletype notification email when an
emergency call is placed.
* A new dedicated emergency bridge webhook can be configured

* Adds new text and html teletype templates
* Adds new teletype module to render template
* Template will distinguish between Emergency and test calls

* Creates a new amqp definition ing `kapi_notifications` for usage
when an emergency call is attempted.
* Publish callback as been added to `cb_notifications`
* Set emergency location data as well as user and device metadata

* Publishes `emergency_bridge` message when stepswitch determines an
emergency call is being attempted
* Adds a new test rules object a resource to define test patterns for
a resource to help differentiate between emergency and test calls.